backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework-docs==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonpickle==3.0.2
packaging==24.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
